ast/visit: include LazyObject in walks
In situations involving PE, it's possible that we walk over an AST that
contains objects taken from the store. With the lazy-objects
optimization, those objects had been neglected: the walker code only
cared for `*object`, not the `Object` interface, for performance reasons.

Now, we'll include the `Object` interface, which covers our lazy
objects, and thus include them in AST walks.

Fixes #5479.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>