Fix tocbot 100vh issue and fix navbar
There seems to be some feature of tocbot which changes the behavior of
the menu based on its size.

I have also fixed the nav as the page scrolls so search is more readily
accessible.

A horizontal scroll bar is also present to allow full built-in table
contents to be viewed on narrow screen sizes. These are currently
clipped.

Signed-off-by: Charlie Egan <charlie@styra.com>