builtins: `json.patch` rework + benchmarks.
This commit contains a reworking of the `json.patch` builtin to use the
new `EditTree` data structure, which makes in-place modification of
nested term structures much more efficient.

This allows a dramatic increase in performance for replace and remove
operations in JSON-Patch sequences, since the term structure now is only
built once, at the end of the sequence of patch applications.

Some basic benchmarks are provided to help illustrate the new
performance properties of the builtin.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>