Load stashed etag from persistence dir on start
This will change the default persistence behaviour. Previously, if OPA
was able to download the bundle from the bundle service then it would do
that even if the bundle hadn't changed. This changes that function to
only download the bundle if the Etag doesn't match. This makes the
functionality similar to the disk storage option.

Signed-off-by: Charlie Egan <charlie@styra.com>