[docs] Use www domain for site base URL
The base URL is used to generate Permalinks in hugo:

https://gohugo.io/content-management/urls/

We then use page permalinks to generate the canonical URL:

https://github.com/open-policy-agent/opa/blob/09f1724270abe9b17a440c035d3837303688e268/docs/website/layouts/partials/meta.html#L47

This causes issues for the algolia crawler since www.openpolicyagent.org
is the domain used in there, and not openpolicyagent.org.

Given that we don't seem to use it either:

```
$ curl -I https://openpolicyagent.org/docs/v0.43.1/policy-language/
HTTP/2 301
location: https://www.openpolicyagent.org/docs/v0.43.1/policy-language/
...
```

I think it makes sense to update this to www. This is being done as
follow on from: https://github.com/open-policy-agent/opa/pull/5706

Signed-off-by: Charlie Egan <charlie@styra.com>