download: copy bundle into buffer only if needed
The need would be if `d.persist` is true: we're writing it into a file
in a later code path.

However, if we don't have bundle persistency enabled, this should allow
us to get rid of one copy of the bundle in memory. To achieve that, the
mechanism used to figure out the bundle size is replaced:

- Before, we'd write it into a buffer that we might not need, and use
  its Len()
- After, we're writing it into a no-op io.Writer that only keeps track
  of the bytes read.

Signed-off-by: Stephan Renatus <stephan@styra.com>