docs: Document pre-release process
This change updates the OPA release process and introduces the
creation of a release candidate prior to the stable release.
The goal is to minimize bugs and regressions in the stable
release and also give users early access to test out features
and bug fixes that will end up in the stable release.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>