runtime: Increase log level for rootless img msg
Since all published OPA images now run with a non-root uid/gid,
there is no need to publish the rootless image tag. Currently
if the rootless variant is run, a log message at the Warn
level is printed.

This change increaes the log level for that message to Error
in order to make it more explict to stop using this variant in the
future.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>