rego: allow for extending the rego targets with plugins (#5939)
The handoff point for plugins right now is the IR: that, the store, and
the open transaction is passed to `PrepareForEval`.

Rego target plugins  aren't plugged in through the plugin manager
machinery, and they don't need to be enabled via configs. They are
more akin to custom builtins, which also become available right
away after having been registered with the ast package.

In the future, another option would be to pass the Wasm bytes: we could
then have a wasmtime-based plugin, and a wazero-based one.

* rego.EvalContext: expose a few fields

* ()*Rego).Compile() is only used by our wasm tests, and doesn't even make
  sense for the "rego" target. Not added to the plugin interface.

* In rego.New, we now only gather plugins once, and have the plugins decide if
  a target is for them or not. This untangles the plugin name from the target
  matching; a plugin can have any name it wants now.

Signed-off-by: Stephan Renatus <stephan@styra.com>
Signed-off-by: Renato Cordeiro <renato@renatocordeiro.com>