testdata: fix malformed json in data.json
This had previously gone unnoticed, but now came up because
the util method used to parse the data.json file is a bit stricter,
it seems.

I don't think that's a bad thing, we shouldn't except malformed
input. Also, I don't think this was malformed by a concious
decision, but more by a copy-paste accident.

Signed-off-by: Stephan Renatus <stephan@styra.com>