rego: expose PrepareOption, add BuiltinFuncs
This is used as a mechanism to inform custom rego target plugins about builtins
that have been defined ad-hoc, via rego.Function1, rego.Function2, etc.

Before, in a setup like this:

```go
	regoArgs = append(regoArgs,
		rego.Target("my_custom_evaluator"),
		// ...
		rego.Function2(builtins.RegalParseModuleMeta, builtins.RegalParseModule),
		rego.Function1(builtins.RegalJSONPrettyMeta, builtins.RegalJSONPretty),
		rego.Function1(builtins.RegalLastMeta, builtins.RegalLast),
	)
```
...the prepared eval state for the custom rego target would not be informed about the
functions defined there.

Now, the PrepareOptions that are passed to the target plugins are used to carry the
information along. To do that, a new field was added, and an accessor function.

In the working stages of this, I had also added accessors for NoInline and PartialEval,
but since the target plugin mechanism explicitly does not support anything PE-related
(yet), I've rolled those changes back.

Signed-off-by: Stephan Renatus <stephan@styra.com>