types: New algorithm for (Any).Union + new benchmarks
This commit swaps out the old algorithm in the (Any).Union method that
would explode on some inputs to one that uses dramatically less memory,
and runs in nearly linear time and memory usage.

This commit also includes improvements to the
`BenchmarkCompileDynamicPolicy` benchmark, as well as new benchmarks for
the `Any` type's `Merge` and `Union` methods.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>