topdown: Improve arith operations by caching small numbers
Currently for arithmetic operations like addition, subtraction etc.
the operands are first converted to the big float type before operating on
them. This conversion can be avoided for small numbers w/o impacting
the accuracy of the result. This change attemtps to avoid the conversion
for small numbers by caching them and thereby helping to speedup some operations.

Fixes: #6021

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>