Update documentation on GCS bundles (#6264)
When an object in GCS contains special characters such as slashes (/) these
need to be url-encoded in the configuration. If not, the bundle will not be found.

e.g. `bundles/bundle.tar.gz` should be entered as `bundles%2fbundle.tar.gz`

This PR adds a small note to help the reader know about this.

Signed-off-by: Dennis Geurts <dennisg@dennisg.nl>