topdown/crypto: Add RawURIs field to JSON certs
This is being added to make it easier to write policy on the contents of
certificate URI SANs. This is where information like SPIFFE IDs etc are
contained and it's helpful to Rego authors to have access to these
values without rebuilding the URI from the parsed data under URIs.

Fixes https://github.com/open-policy-agent/opa/issues/6416

Signed-off-by: Charlie Egan <charlie@styra.com>