.github/workflows: Update name of the checkout action token
This PR attempts to fix the failures seen in the Post Merge workflow.
Previously the token used in the checkout action had probably expired
and a new token was generated to test out this hypothesis. Now that
the Post Merge workflow succeeds, we're renaming the token as before
and updating the token value in the repo.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>