Fixing race issue in test logger
`WithFields()` would create a copy of the logger, but with a new mutex instance.
This would result in multiple logger copies concurrently modifying the `entries` field,
but within the scope of different locks.

Concretely, this would mean the test creating the logger and calling `Entries()` would not be using the same lock
as the code under test calling `append()`.

Signed-off-by: Johan Fylling <johan.dev@fylling.se>