docs: update version for non docs pages
This sets the default version for all non docs pages to latest.

We have recently started to list some pages in the search index from non
docs pages, when a user makes a search, they search within the current
version. This will make pages from the rest of the site available in the
index when searching from 'latest'.

Signed-off-by: Charlie Egan <charlie@styra.com>