docs: Fix issue in versioning docs for search
https://github.com/open-policy-agent/opa/pull/6526 contained two errors,
that I can see. First, it used `Permalink` rather than `RelPermalink`,
this contains the host and scheme on build and so the older versions of
pages were all getting version set to the latest.

Even if that had worked, the setting of version was incorrect and set a
locally scoped var instead of the variable for the whole file.

Signed-off-by: Charlie Egan <charlie@styra.com>