sdk: fix bug while activating the bundle plugin
For some reason the bundle plugin activation with multiple optimized
bundles, builded in v1 syntax, results in parsing error.

During the activation of the bundle, when we need to parse the rego
modules, the ParserOptions is not forwarded to the `bundle.ActivateOps`,
so the v1 syntax does not seem to be properly parsed.

Signed-off-by: Francisco Rodrigues <ednofco@gmail.com>