logs: add ability to enrich decision logs
With the ability to add arbitrary data to decision logs, we give OPA
users opportunity to enrich logs with any context or dynamic data
fetched from custom built-ins.

Also, we provide a building block to make it easier for internal
built-ins to communicate with the decision log plugin.

This data gets appended to an `extra` key in the decision log, that may
contain any json data.

Signed-off-by: Francisco Rodrigues <ednofco@gmail.com>